split naive functionality into modules
