+++ src/column.rs
@@ -1,0 +1,7 @@
+use datatypes::DataType;
+
+#[derive(Debug, RustcEncodable, RustcDecodable)]
+pub struct Column {
+    pub name: String,
+    pub datatype: DataType,
+}
