+++ src/lib.rs
@@ -1,0 +1,11 @@
+#[macro_use]
+extern crate quick_error;
+extern crate bincode;
+extern crate rustc_serialize;
+
+pub mod column;
+pub mod datatypes;
+pub mod result;
+pub mod errors;
+pub mod row;
+pub mod table;
