+++ src/result.rs
@@ -1,0 +1,3 @@
+use errors::ConstableError;
+
+pub type ConstableResult<T> = Result<T, ConstableError>;
